requirements.txt
doit==0.30.3
panflute==1.10.6
PyYAML==3.12
